Deleted the Camera visit methods in favor of the one in the prepass traversal
